+++ src/api/fast_api.py
@@ -1,9 +1,12 @@
-from fastapi import FastAPI, HTTPException, Request
-from fastapi.responses import HTMLResponse
+from fastapi import APIRouter, FastAPI, HTTPException, Request
+from fastapi.responses import HTMLResponse, JSONResponse
 from pydantic import BaseModel
-from . import __version__
-from utils.constants import DEFAULT_GREETING
+from starlette.exceptions import HTTPException as StarletteHTTPException
+
+from utils.constants import API_PREFIX, GREETING
 from utils.helper import normalize_name
+
+from . import __version__
 
 app = FastAPI(
     title="sample API",
@@ -19,6 +22,11 @@ app = FastAPI(
     },
 )
 
+api_v1 = APIRouter(
+    prefix=API_PREFIX,
+    tags=["V1"],
+)
+
 
 class GreetRequest(BaseModel):
     name: str
@@ -28,12 +36,12 @@ class GreetResponse(BaseModel):
     message: str
 
 
-@app.get("/version")
+@app.get("/version", tags=["Version"])
 def version():
     return {"version": app.version}
 
 
-@app.get("/", response_class=HTMLResponse)
+@app.get("/", response_class=HTMLResponse, tags=["Home"])
 async def read_root(request: Request):
     return """
     <html>
@@ -75,19 +83,52 @@ async def read_root(request: Request):
     """
 
 
-@app.get("/health")
+@app.get("/health", tags=["Help"])
 def health_check():
     return {"status": "ok"}
 
 
-@app.post("/greet", response_model=GreetResponse)
+@api_v1.get("/help", tags=["Help"])
+def get_help():
+    return JSONResponse(
+        status_code=200,
+        content={
+            "message": "Welcome to the Sample BoilerPlate API! Visit /docs for API documentation."
+        },
+    )
+
+
+@api_v1.post("/greet", response_model=GreetResponse)
 def greet_user(payload: GreetRequest):
     clean_name = normalize_name(payload.name)
 
     if not clean_name:
         raise HTTPException(status_code=400, detail="Invalid name provided")
 
-    return {"message": f"{DEFAULT_GREETING}, {clean_name} 👋"}
+    return {"message": f"{GREETING}, {clean_name} 👋"}
+
+
+@app.exception_handler(StarletteHTTPException)
+async def http_exception_handler(request: Request, exc: StarletteHTTPException):
+    if exc.status_code == 404:
+        endpoint = request.url.path
+
+        return JSONResponse(
+            status_code=404,
+            content={
+                "error": f"Endpoint '{endpoint}' does not exist, use /docs to see available endpoints.",
+                "status": 404,
+            },
+        )
+
+    # fallback for other HTTP errors
+    return JSONResponse(
+        status_code=exc.status_code,
+        content={"error": exc.detail},
+    )
+
+
+app.include_router(api_v1)
 
 
 def start():