+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 All notable changes to this repository will be documented in this file.
 
-## [1.0.0]
+## [1.0.0] Sat,10 jan 2026
 
 - Initial release.
-- Include boilerplate with streamlit and fast api
+- Include boilerplate without streamlit and fast api
+
+## [1.1.0] Sat,17 Jan 2026
+
+- Updated UI.
+- Add clean script.
+- Add exception handling on API.